Feature: salvando os dados do cartão por mensageria.

+++ ms-cartoes/src/main/kotlin/com/jfb/mscartoes/adapter/in/web/response/CartoesPorClienteResponse.kt
@@ -5,5 +5,5 @@ import java.math.BigDecimal
 data class CartoesPorClienteResponse(
   val nome: String,
   val bandeira: String,
-  val limiteLiberado: BigDecimal
+  val limiteLiberado: Long
 )
